pyproject.toml
[build-system]
requires = ["uv_build>=0.8.4,<0.9.0"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-name = "src"
module-root = ""

[project]
name = "veb_centor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "asyncpg>=0.31.0",
    "fastapi>=0.132.0",
    "greenlet>=3.3.2",
    "jieba>=0.42.1",
    "langchain-community>=0.4.1",
    "langchain-text-splitters>=1.1.1",
    "openai>=2.22.0",
    "pgvector>=0.4.2",
    "psycopg[binary]>=3.3.3",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "pypdf>=6.7.2",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "sqlalchemy>=2.0.46",
    "sqlalchemy-utils>=0.42.1",
    "uvicorn>=0.41.0",
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
]

[tool.pytest.ini_options]
addopts = "-v -s"
testpaths = ["src/tests"]
